Started working on the renderer and the Ghoul II routines. A server-side renderer is required in order to make the Ghoul II API calls properly work. Instead of taking an existing renderer, the goal is to port the renderer from the ground up so that we only include the routines we actually need to make server-side rendering work.
The Ghoul II routines are residing in the code/ghoul2 directory, next to the
shared Ghoul II header files. The renderer routines reside in the
code/rd-dedicated directory. Unlike the original renderer, the dedicated
renderer is not included as a module since SoF2Plus will be dedicated only.

Note that this is a work in progress and expect breakage for a some time to
come.