Added map cycle support. The same structure is maintained as the original SoF2MP. The map cycle file must contain a GP2 structure with a base group called "mapcycle" which contains 1..n subgroups. If these conditions are not met, or the map cycle file is not found, the command will error out and either restart the map in case of a running server or do nothing if the server is not running (yet).
Each subgroup must contain a "command" pair and optionally a "cvars" subgroup.
The "cvars" subgroup contains pairs of CVARs to be set.

The next map subgroup is not checked by the server: it is the server owner their
responsibility to at least provide a valid command to switch the server to the
next map. Failing to do so usually results in a game continuing even after the
set score- or timelimit condition is met.

Users can set the "sv_mapcycle" CVAR to any file containing a valid
map cycle that can be parsed by SoF2Plus. The appropriate file extension
(.mapcycle) is recommended for use but not enforced in any way.