Added clipmap sub-BSP support. Sub-BSPs can now be loaded properly. Since both the map/main BSP (CM_LoadMap function) and any sub-BSP (CM_LoadSubBSP function) have very similar initialization routines, a function is added which loads the actual BSP contents, named CM_LoadBSPFile, to minimize the amount of duplicate code in the aforementioned functions.
The "clientload" function parameter is also removed from CM_LoadMap, since this
is always false in SoF2Plus.

The clipmap tracing/testing code is also updated and now aware of which clipmap
BSP a model belongs to and traces accordingly.

This commit also reverts the change to the "cm" pointer in commit bc4733e.
Instead of pointing at the active BSP clipmap, which basically affects all
clipmap code, the "cm" variable is removed and a new global variable named "cmg"
is added instead. The "cmg" variable always points to the main BSP, similar to
how the old "cm" instance worked.

The downside to this implementation is that a lot of code changing is required,
as shown in this commit. The original idea of having "cm" point to the active
BSP clipmap almost only affects the tracing code, which may need access to both
the main BSP and a sub-BSP in one trace. This involves switching the active
clipmap quite often, making the use of "cm" ambiguous to the programmer across
traces.

Adding the new "cmg" variable makes the code distinct and less prone to error.
A "cm" pointer is now always a local variable pointing to any clipmap BSP.