Fix clipping issue (#2)
In SoF2, there is an additional check in the case where start of the trace is equal to the end of the trace (effectively a position test). There's another check done on the half-extents. If the half-extents are 0, then it moves on to sweeping through the world.

Adding the half extents check effectively solves the clipping into walls issue when you're crouch jumping and releasing crouch at the apex of the jump.